add create_chds.pl
this is basically a simple loop over MAME's chdman tool that lets you
create more than one chd with a single command.  It also allows you to
specify either the path to the .gdi itself or the path to the directory
with the .gdi in it, and IDK if chdman lets you do that.